backend/pyproject.toml
[project]
name = "minimax-music-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "mutagen>=1.47.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "uvicorn[standard]>=0.44.0",
]
